% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asocnom.R
\name{asocnom}
\alias{asocnom}
\title{Medida de asociación Lambda de Goodman Kruskal.}
\usage{
asocnom(X, Y, conf, nombrex, nombrey)
}
\arguments{
\item{X}{vector numérico.}

\item{Y}{vector numérico.}

\item{Conf}{Valor que indica el nivel del confianza para crear el intervalo (0<conf<1).}

\item{Nombrex}{cadena de caracteres que indica nombre de la variable X.}

\item{Nombrey}{cadena de caracteres que indica nombre de la variable Y.}
}
\description{
devuelve un mensaje con el coeficiente Lambda simétrico para variables nominales.
 Además de un intervalo con un nivel de confianza dado.
}
