% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlacion.R
\name{correlacion}
\alias{correlacion}
\title{Coeficiente de correlación}
\usage{
correlacion(X, Y, nombres, método)
}
\arguments{
\item{X:}{vector numérico.}

\item{Y:}{vector numérico.}

\item{Nombres:}{vector de caracteres con las preguntas o afirmaciones
de las variables (el número de columnas de data debe ser igual a p).}

\item{Método}{: cadena de caracteres indicando qué coeficiente de
correlación se debe calcular. "pearson", "kendall" o "spearman".}
}
\description{
Devuelve un mensaje con el valor del coeficiente
 de correlación lineal en función del metodo indicado. Pearson,
  Kendall o Spearman.
}
