% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.cuan.2cual.R
\name{desc.cuan.2cual}
\alias{desc.cuan.2cual}
\title{Tabla de resúmenes numéricos de una variable cuantitativa respecto de dos cualitativas}
\usage{
desc.cuan.2cual(
  datos,
  X1,
  X2,
  X3,
  etiquetas2,
  etiquetas3,
  nombre1,
  nombre2,
  nombre3,
  título
)
}
\arguments{
\item{X1}{vector de respuestas de la variable cuantitatva (tamaño n)}

\item{X2}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{X3}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{Datos}{data.frame con los datos a analizar.}

\item{Etiquetasx2}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 2}

\item{Etiquetasx3}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 3}

\item{Nombre1}{cadena de caracteres indicando la pregunta realizada para la variable x1 (no admite espacios)}

\item{Nombre2}{cadena de caracteres indicando la pregunta realizada para la variable x2 (no admite espacios)}

\item{Nombre3}{cadena de caracteres indicando la pregunta realizada para la variable x3 (no admite espacios)}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la tabla}
}
\description{
Esta funcion devuelve la tabla de resúmenes numéricos básicos de la variable
 cuantitativa respecto de la variable cualitativa, incluyendo el tamaño muestral, media,
  mediana, desviación estandar e intervalo de confianza al 95% de X1 respecto de X2 y X3
}
