% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.cuan.cual.R
\name{desc.cuan.cual}
\alias{desc.cuan.cual}
\title{Tabla de resúmenes numéricos de la variable cuantitativa respecto de la cualitativa.}
\usage{
desc.cuan.cual(X1, X2, etiquetasX2, nombreX1, nombreX2, título)
}
\arguments{
\item{X1}{vector de respuestas de la variable cuantitatva (tamaño n)}

\item{X2}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{Datos}{data.frame con los datos a analizar.}

\item{Etiquetasx2}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 2}

\item{Nombrex1}{cadena de caracteres indicando la pregunta realizada para la variable x1 (no admite espacios)}

\item{Nombrex2}{cadena de caracteres indicando la pregunta realizada para la variable x2 (no admite espacios)}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la salida}
}
\description{
Esta función devuelve la tabla de resúmenes numéricos básicos de la variable
 cuantitativa respecto de la variable cualitativa, incluyendo el tamaño muestral, media,
  mediana, desviación estandar e intervalo de confianza al 95% de X1 respecto de X2
}
