% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.cuan.subg.R
\name{desc.cuan.subg}
\alias{desc.cuan.subg}
\title{Tabla de resúmenes numéricos para una variable cuantitava dividida en dos partes.}
\usage{
desc.cuan.subg(datos, x, nombre, valor, título)
}
\arguments{
\item{Datos}{data.frame con los datos a analizar.}

\item{X}{vector de respuestas de la variable a analizar (tamaño n)}

\item{Nombre}{cadena de caracteres indicando la pregunta realizada. ( No admite espacios)}

\item{Título}{cadena de caracteres que indica el nombre que se le quiere dar a la tabla}

\item{Valor}{de tipo númerico, valor dado para separar nuestra variable en dos partes.}
}
\description{
Esta función devuelve una tabla de resúmenes númericos básicos, tamaño muestral,
 media, mediana, desviación estandar e intervalo de confianza al 95% de una variable cuantitativa
 dada dividida en dos partes, la primera menor que un valor y la segunda mayor que ese mismo valor dado.
}
