% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.cual.2cual .R
\name{freq.cual.2cual}
\alias{freq.cual.2cual}
\title{Tabla de frecuencias de tresz variables cualitativas}
\usage{
freq.cual.2cual(
  datos,
  X1,
  X2,
  X3,
  etiquetasx1,
  etiquetasx2,
  etiquetasx3,
  nombre1,
  nombre2,
  nombre3,
  título
)
}
\arguments{
\item{X1}{vector de respuestas de la variable cualitativa 1 (filas)(tamaño n)}

\item{X2}{vector de respuestas de la variable cualitativa 3 (columnas)(tamaño n)}

\item{etiquetasx1}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 1}

\item{etiquetasx2}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 2}

\item{etiquetasx3}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 3}

\item{Datos}{data.frame con los datos a analizar.}

\item{Nombre1}{cadena de caracteres indicando la pregunta realizada para la variable x1}

\item{Nombre2}{cadena de caracteres indicando la pregunta realizada para la variable x2}

\item{Nombre3}{cadena de caracteres indicando la pregunta realizada para la variable x3}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la salida}
}
\description{
Función que genera una tabla de frecuencias absolutas
de las variable x1, x2 y x3.
}
