% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.dic.R
\name{freq.dic}
\alias{freq.dic}
\title{Tabla de frecuencias para variables binarias o dicotómicas.}
\usage{
freq.dic(datos, X, etiquetas, titulo)
}
\arguments{
\item{X}{vector de respuestas de la variable dicotómica a analizar (tamaño n)}

\item{Datos}{data.frame con los datos a analizar.}

\item{Etiquetas}{vector de cadena de caracteres con las posibles respuestas.}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la tabla}
}
\description{
Función que genera una tabla de frecuencias absolutas, relativas y acumuladas
para la variable x.
}
