% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.nom.R
\name{freq.nom}
\alias{freq.nom}
\title{Tabla de frecuencias para variables nominales}
\usage{
freq.nom(datos, X, etiquetas, título)
}
\arguments{
\item{X}{vector de respuestas de la variable nominal a analizar (tamaño n)}

\item{Datos}{data.frame con los datos a analizar.}

\item{Etiquetas}{vector de cadena de caracteres con las posibles respuestas.}

\item{Título}{cadena de caracteres que indica el nombre que se le quiere dar a la tabla}
}
\value{

}
\description{
Función que genera una tabla de frecuencias absolutas y relativas junto con sus acumuladas
para la variable x.
}
