% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.ord.R
\name{freq.ord}
\alias{freq.ord}
\title{Tabla de frecuencias para variables ordinales}
\usage{
freq.ord(datos, X, etiquetas, título)
}
\arguments{
\item{datos}{data.frame con los datos a analizar.}

\item{X}{vector de respuestas de la variable ordinal a analizar (tamaño n)}

\item{Etiquetas}{vector de cadena de caracteres con las posibles respuestas.}

\item{Título}{cadena de caracteres que indica el nombre que se le quiere dar a la tabla}
}
\value{

}
\description{
Función que genera una tabla de frecuencias absolutas y relativas junto con sus acumuladas
de la variable x.
}
