% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.2cont.2nom.R
\name{graf.2cont.2nom}
\alias{graf.2cont.2nom}
\title{Visualización gráfica de dos variables continuas clasificadas por dos variables nominales}
\usage{
graf.2cont.2nom(
  datos,
  X1,
  X2,
  X3,
  X4,
  nombre1,
  nombre2,
  nombre3,
  nombre4,
  etiquetas1,
  etiquetas2,
  título
)
}
\arguments{
\item{X1}{vector de respuestas de la variable cuantitativa 1 (columnas) (tamaño n)}

\item{X2}{vector de respuestas de la variable cuantitativa 2 (filas) (tamaño n)}

\item{X3}{vector de respuestas de la variable cualitativa 3 (tamaño n)}

\item{X4}{vector de respuestas de la variable cualitativa 4 (tamaño n)}

\item{Datos}{data.frame con los datos a analizar.}

\item{Etiquetasx3}{vector de caracteres con las posibles respuestas de la variable 3}

\item{Etiquetasx4}{vector de caracteres con las posibles respuestas de la variable 4}

\item{Nombre1}{cadena de caracteres con el nombre de la variable 1}

\item{Nombre2}{cadena de caracteres con el nombre de la variable 2}

\item{Nombre3}{cadena de caracteres con el nombre de la variable 3}

\item{Nombre4}{cadena de caracteres con el nombre de la variable 4}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar al gráfico}
}
\description{
Esta función relaciona dos variables continuas según dos variables nominales , cada una por separado
}
