% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.2cont.nom.R
\name{graf.2cont.nom}
\alias{graf.2cont.nom}
\title{Visualización gráfica Relación entre dos variables continuas clasificada por una nominal.}
\usage{
graf.2cont.nom(datos, X1, X2, X3, nombre1, nombre2, nombre3, etiquetas, título)
}
\arguments{
\item{datos}{data.frame con los datos a analizar.}

\item{X1}{vector de respuestas de la variable continua1 (columnas) (tamaño n)}

\item{X2}{vector de respuestas de la variable continua 2 (filas) (tamaño n)}

\item{X3}{vector de respuestas de la variable nominal (tamaño n)}

\item{nombre3}{cadena de caracteres con el nombre de la variable 3}

\item{Etiquetas}{vector de caracteres con las posibles respuestas de la variable 3}

\item{Nombre1}{cadena de caracteres con el nombre de la variable 1}

\item{Nombre2}{cadena de caracteres con el nombre de la variable 2}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar al gráfico}
}
\description{
Esta función relaciona dos variables continuas según una variable nominal,
muestra los gráficos de dispersión por separado.
}
