% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.cual.cont.R
\name{graf.cont.cual}
\alias{graf.cont.cual}
\title{Visualización gráfica variable continua en función de una cualitativa.}
\usage{
graf.cont.cual(X1, X2, etiquetasx1, nombrex1, nombrex2, título)
}
\arguments{
\item{X1}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{X2}{vector de respuestas de la variable continua (tamaño n)}

\item{Datos}{data.frame con los datos a analizar.}

\item{EtiquetasX1}{vector de cadena de caracteres con las posibles respuestas de la variable 1}

\item{Nombrex1}{cadena de caracteres indicando la pregunta realizada para la variable x1}

\item{Nombrex2}{cadena de caracteres indicando la pregunta realizada para la variable x2}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a los gráficos}
}
\description{
Esta función genera el histograma, gráfico de densidad ,
caja y bigote y tipo jitter de la variable continua en función de una cualitativa
}
