% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.cont.R
\name{graf.cont}
\alias{graf.cont}
\title{Visualización gráfica de variables continuas.}
\usage{
graf.cont(datos, X, nombre, título)
}
\arguments{
\item{X}{vector de respuestas de la variable continua  (tamaño n)}

\item{Datos}{data.frame donde se encuentra X}

\item{Nombre}{cadena de caracteres indicando la pregunta realizada o nombre de X}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la salida}
}
\value{
Diagrama de caja, histograma con media e histograma más densidad
}
\description{
Función que genera un diagrama de caja, histograma con media e
 histograma más densidad de la variable X.
}
