% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.cor.R
\name{graf.cor}
\alias{graf.cor}
\title{Visualización gráfica matriz de correlaciones}
\usage{
graf.cor(data, nombres, título, método)
}
\arguments{
\item{Data:}{data.frame formada por las p variables que se desea estudiar su correlación.}

\item{Nombres:}{vector de caracteres con los nombres de las variables (su longitud igual a p).}

\item{Título}{: cadena de caracteres indicando el nombre que se le quiere dar a la salida.}

\item{Método}{: cadena de caracteres indicando qué coeficiente de correlación se debe calcular. "pearson", "kendall" o "spearman".}
}
\description{
La función graf.cor genera un gráfico muy visual de las correlaciones entre varias variables.
 Se muestra la matriz con los coeficientes de correlación. En la diagonal están las variables, por encima
 están unos círculos de colores, entre más intensidad del color, ya sea azul o rojo, mayor es la correlación,
 colores ténues significan correlación baja; el tamaño de los círculos está asociado al valor absoluto de correlación.
 Por debajo de la diagonal se observan los valores exactos de correlación en colores
}
