% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.cual.cual.R
\name{graf.cual.cual}
\alias{graf.cual.cual}
\title{Visualización gráfica de dos variables cualitativas.}
\usage{
graf.cual.cual(
  datos,
  x1,
  x2,
  etiquetasx1,
  etiquetasx2,
  nombrex1,
  nombrex2,
  título
)
}
\arguments{
\item{Datos}{data.frame con los datos a analizar.}

\item{X1}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{X2}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{Etiquetasx1}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 1}

\item{Etiquetasx2}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 2}

\item{Nombrex1}{cadena de caracteres indicando la pregunta realizada para la variable x1}

\item{Nombrex2}{cadena de caracteres indicando la pregunta realizada para la variable x2}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a los gráficos}
}
\description{
Esta funcion devuelve un gráfico de barras para
 observar la relación que existe entre dos variables cualitativas.
}
