% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.cual.R
\name{graf.cual}
\alias{graf.cual}
\title{Visualización gráfica para variables cualitativas}
\usage{
graf.cual(datos, x, etiquetas, nombre, título)
}
\arguments{
\item{Datos}{data.frame donde se encuentra X}

\item{X}{vector de respuestas de la variable cualitativas a analizar (tamaño n)}

\item{Etiquetas}{vector de cadena de caracteres con las posibles respuestas no numéricas}

\item{Nombre}{cadena de caracteres indicando la pregunta realizada o nombre de X}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la salida}
}
\value{
Diagrama de barras y de sectores
}
\description{
Función que genera un diagrama de barras y diagrama de sectores.
}
