% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.cuan.cuan.R
\name{graf.cuan.cuan}
\alias{graf.cuan.cuan}
\title{Visualización gráfica de una variable continua respecto de otra continua}
\usage{
graf.cuan.cuan(X1, X2, nombrex1, nombrex2, título)
}
\arguments{
\item{X1}{vector de respuestas de la variable continua 1 (tamaño n)}

\item{X2}{vector de respuestas de la variable continua 2 (tamaño n)}

\item{datos}{data.frame con los datos a analizar.}

\item{Nombrex1}{cadena de caracteres indicando la pregunta realizada para la variable x1}

\item{Nombrex2}{cadena de caracteres indicando la pregunta realizada para la variable x2}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar al gráfico}
}
\description{
La siguiente función muestra el gráfico de dispersión entre dos variables continuas dadas.
}
