% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.disc.R
\name{graf.disc}
\alias{graf.disc}
\title{Visualización gráfica de variables discretas}
\usage{
graf.disc(datos, X, nombre, título)
}
\arguments{
\item{X}{vector de respuestas de la variable discreta a analizar (tamaño n)}

\item{Datos}{data.frame donde se encuentra X}

\item{Nombre}{cadena de caracteres indicando la pregunta realizada o nombre de X}

\item{Título}{cadena de caracteres indicando el nombre que se le quiere dar a la salida}
}
\value{
Diagrama de cajas y de densidad
}
\description{
Función que genera un diagrama de caja y gráfico de densidad de la variable X
}
