% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.val1.R
\name{graf.val1}
\alias{graf.val1}
\title{Visualización gráfica variables de valoración (Escala Likert)}
\usage{
graf.val1(datos, x, etiquetas, nombre)
}
\arguments{
\item{Datos}{data.frame donde se encuentra X}

\item{X}{vector de respuestas de la variable cualitativas a analizar (tamaño n)}

\item{Etiquetas}{vector de cadena de caracteres con las posibles respuestas no numéricas}

\item{Nombre}{cadena de caracteres indicando la pregunta realizada o nombre de X}
}
\value{
Diagrama de barras, mapa de calor y gráfico de densidad.
}
\description{
Función que genera un diagrama de barras, mapa de calor y gráfico de densidad de la variable X .
}
