% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.val3.cual.R
\name{graf.val3.cual}
\alias{graf.val3.cual}
\title{Visualización gráfica variables de valoración (Escala Likert) para bloques de tres con
 las mismas etiquetas en función de una variable cualitativa}
\usage{
graf.val3.cual(
  datos,
  x1,
  x2,
  x3,
  x4,
  etiquetas,
  etiquetasx4,
  preguntas,
  nombrex4
)
}
\arguments{
\item{datos}{data.frame con los datos a analizar.}

\item{X1}{vector de respuestas de la variable de valoración (tamaño n)}

\item{X2}{vector de respuestas de la variable de valoración (tamaño n)}

\item{X3}{vector de respuestas de la variable de valoración (tamaño n)}

\item{X4}{vector de respuestas de la variable cualitativa (tamaño n)}

\item{Etiquetas}{vector de cadena de caracteres con las posibles respuestas  de las variables de valoración}

\item{Etiquetasx4}{vector de cadena de caracteres con las posibles respuestas no numéricas de la variable 4}

\item{Preguntas}{vector de cadenas de caracteres con las preguntas o nombres en orden de x1, x2 y x3}

\item{Nombrex4}{cadena de caracteres indicando la pregunta realizada para la variable x4}
}
\description{
Función que genera un diagrama de barras y gráfico de densidad para una fácil
 visualización y comparación de 3 variables de tipo ordinal en función de otra variable cualitativa
}
