% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf.val.3.R
\name{graf.val3}
\alias{graf.val3}
\title{Visualización gráfica para variables de valoración (Escala Likert) en bloques de tres.}
\usage{
graf.val3(datos, x1, x2, x3, etiquetas, preguntas)
}
\arguments{
\item{Datos}{data.frame donde se encuentra X}

\item{X1}{vector de respuestas de la variable 1 (tamaño n)}

\item{X2}{vector de respuestas de la variable 2 (tamaño n)}

\item{X3}{vector de respuestas de la variable 3 (tamaño n)}

\item{Etiquetas}{vector de cadenas de caracteres con las posibles respuestas no numéricas}

\item{Preguntas}{vector de cadenas de caracteres con las preguntas o afirmaciones en orden de x1, x2 y x3}
}
\value{
Diagrama de barras, mapa de calor y gráfico de densidad.
}
\description{
Función que genera un diagrama de barras para una fácil
 visualización y comparación de tres variables de tipo ordinal con las mismas etiquetas.
}
