% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz.cor.R
\name{matriz.cor}
\alias{matriz.cor}
\title{Matriz de correlaciones}
\usage{
matriz.cor(data, nombres, título, método)
}
\arguments{
\item{Data:}{data.frame formada por las p variables que se desea estudiar
su correlación.}

\item{Nombres:}{vector de caracteres con las preguntas o nombres de las variables (el número de columnas de data debe ser igual a p).}

\item{Título}{: cadena de caracteres indicando el nombre que se le quiere dar a la salida.}

\item{Método}{: cadena de caracteres indicando qué coeficiente de correlación
se debe calcular. "pearson", "kendall" o "spearman".}
}
\description{
Esta función genera una matriz con las correlaciones
 dos a dos de hasta p variables en función del método indicado.
}
