% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablaLatex.R
\name{tablaLatex}
\alias{tablaLatex}
\title{Tablas en Latex}
\usage{
tablaLatex(tabla, caption = NULL)
}
\arguments{
\item{Tabla}{tabla en R que se desea escribir en Latex}

\item{Caption}{título de la tabla, por defecto nulo}
}
\description{
Función para imprimir tablas en latex
}
\references{
\url{http://gauss.inf.um.es/8jornadasR/filestaller/T1_descriptivos.html}
}
